Added Create Table functionality
We have added the `Create table` functionality on the endpoint `/create`.
This new action allows the user to create a new table with the following params:
 - Table name
 - Column names

These new table files will be added inside the `tables` folder with the following name: `tableName.json` where tableName is the name provided by the user.

Also added `jackson` dependency to the pom file, so we can manage JSON objects and files properly.

Will be needed a cleaning to the repo in another PullRequest.

Updated `.gitignore`.